+++ packages/browser-wallet/src/popup/pages/Account/i18n/da.ts
@@ -47,6 +47,7 @@ const t: typeof en = {
         },
     },
     confirmTransfer: {
+        submitted: 'Transaktionen er blevet indsendt!',
         buttons: {
             back: 'tilbage',
             send: 'Send',
@@ -337,12 +338,6 @@ const t: typeof en = {
             lowerBakerStake:
                 'Du er ved at indsende en transaktion som formindsker din baker stake. At formindske din baker stake har en cool-down periode, hvilket betyder at ændringen ikke træder i kraft med det samme.\n\nBakeren kan ikke fjernes og din baker stake kan ikke ændres indtil cool-down perioden er overstået.',
             removeBaker: 'Er du sikker du vil lave denne transaktion som stopper baking?',
-            notice: {
-                start: 'Når din transaktion er finalized, vil baker registreringen starte fra den efterfølgende pay day.\n\nHusk at genstarte din node med de nye baker keys',
-                update: 'Når din transaktion er blevet finalized, vil baker opdateringen træde i kraft fra den efterfølgende pay day.',
-                updateKeys:
-                    'Når din transaktion er finalized, vil de nye baker keys blive gyldige fra den efterfølgende pay day.\n\nDu bør derfor genstarte din node med de nye baker keys så tæt på den næste pay day som muligt.',
-            },
         },
         configureDelegation: {
             register:
@@ -350,10 +345,18 @@ const t: typeof en = {
             lowerDelegationStake:
                 'Du er ved at indsende en transaktion som formindsker din delegation stake. Det vil træde i kraft efter {{ cooldownPeriod }} dage og din delegation saldo kan ikke ændres i denne periode.',
             remove: 'Er du sikker du vil fjerne din delegation?',
-            notice: {
-                start: 'Når din transaktion er blevet finalized, vil delegation starte fra den efterfølgende pay day.',
-                update: 'Når din transaktion er blevet finalized, vil delegation opdateringen træde i kraft fra den efterfølgende pay day.',
-            },
+        },
+    },
+    transactionPopup: {
+        configureBaker: {
+            start: 'Når din transaktion er finalized, vil baker registreringen starte fra den efterfølgende pay day.\n\nHusk at genstarte din node med de nye baker keys',
+            update: 'Når din transaktion er blevet finalized, vil baker opdateringen træde i kraft fra den efterfølgende pay day.',
+            updateKeys:
+                'Når din transaktion er finalized, vil de nye baker keys blive gyldige fra den efterfølgende pay day.\n\nDu bør derfor genstarte din node med de nye baker keys så tæt på den næste pay day som muligt.',
+        },
+        configureDelegation: {
+            start: 'Når din transaktion er blevet finalized, vil delegation starte fra den efterfølgende pay day.',
+            update: 'Når din transaktion er blevet finalized, vil delegation opdateringen træde i kraft fra den efterfølgende pay day.',
         },
     },
     accountPending: 'Denne konto er stadig ved at blive oprettet.',